+++ src/impl/utils/setup-services.ts
@@ -1,6 +1,7 @@
 import { Connection } from "../../sqlite3";
 import { TableDefinition } from "../../undo-log-tables";
-import { Utils } from "../../utilities";
+import { DatabaseDefinitionServices } from "../../utils/database-definition-services";
+import { DatabaseUtilitiesServices } from "../../utils/database-utilities-services";
 import { PrivateServices } from "../../utils/private-services";
 import { SetupServices } from "../../utils/setup-services";
 import { UndoLogUtilityServices } from "../../utils/undo-log-utility-services";
@@ -9,11 +10,13 @@ import { UndoLogUtilityServices } from "../../utils/undo-log-utility-services";
 export class SetupServicesImpl implements SetupServices {
   private connection: Connection;
   private prefix: string;
-  private utils: Utils;
+  private definitions: DatabaseDefinitionServices;
+  private utils: DatabaseUtilitiesServices;
   constructor(srv: UndoLogUtilityServices & PrivateServices) {
     this.connection = srv.connection;
     this.prefix = srv.prefix;
-    this.utils = srv.utils;
+    this.definitions = srv.databases.definition;
+    this.utils =  srv.databases.utils;
   }
 
   async updateUndoLogTable<T extends Record<string, any> & { id: number; }>(
@@ -28,9 +31,9 @@ export class SetupServicesImpl implements SetupServices {
   }
 
   async createUndoLogTable(tableName: string, definition: TableDefinition) {
-    const columns = this.utils.createColumnDefinitions(definition);
-    const foreigns = this.utils.createForeignKeys(definition, this.prefix);
-    const uniques = this.utils.createUniqueKeys(definition);
+    const columns = this.definitions.createColumnDefinitions(definition);
+    const foreigns = this.definitions.createForeignKeys(definition, this.prefix);
+    const uniques = this.definitions.createUniqueKeys(definition);
     const query = `CREATE TABLE ${this.prefix}${tableName} (${columns}${foreigns}${uniques});`;
     await this.connection.execute(query);
   }
