+++ src/utils/database-query-services.ts
@@ -1,0 +1,8 @@
+
+export interface DatabaseQueryServices {
+  hasTableId(name: string, id: number): Promise<boolean>;
+  tableHas<T extends Record<string, any> & { id: number; }>(
+    tableName: string,
+    row: T
+  ): Promise<[boolean, string[]]>;
+}
